refactor: 모든 View에 대해, ViewModel 의존성 주입으로 변경 (#6)
- 위치 변경
